chore: Move common yaml util
